Filter/requirements.txt
asgiref==3.3.1
Babel==2.9.0
Django==3.1.6
django-filter==2.4.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.18
pytz==2021.1
sqlparse==0.4.1
